requirements.txt
streamlit
google-generativeai
Pillow
python-dotenv
PyPDF2
numpy
pandas